refactor WebDriver module
