+++ Sources/WebDriver/WebDriverClient.swift
@@ -15,31 +15,11 @@
 import Foundation
 
 #if canImport(FoundationNetworking)
-  import FoundationNetworking
-
-  /// Until we get "async" implementations of URLSession in corelibs-foundation, we use our own polyfill.
-  extension URLSession {
-    public func data(for request: URLRequest) async throws -> (Data, URLResponse) {
-      return try await withCheckedThrowingContinuation { continuation in
-        let task = self.dataTask(with: request) { (data, response, error) in
-          guard let data = data, let response = response else {
-            let error = error ?? URLError(.badServerResponse)
-            return continuation.resume(throwing: error)
-          }
-          continuation.resume(returning: (data, response))
-        }
-        task.resume()
-      }
-    }
-  }
+import FoundationNetworking
 #endif
 
-public enum WebDriverError: Error {
-  case httpError(String)
-}
-
 public struct WebDriverClient {
-  private let client: WebDriverHTTPClient
+  private let client: any WebDriverHTTPClient
   let driverEndpoint: URL
   let sessionId: String
 
@@ -71,8 +51,9 @@ public struct WebDriverClient {
     """#
 
   public static func newSession(
-    endpoint: URL, body: String = defaultSessionRequestBody,
-    httpClient: URLSession
+    endpoint: URL, 
+    body: String = defaultSessionRequestBody,
+    httpClient: (any WebDriverHTTPClient)? = nil
   ) async throws -> WebDriverClient {
     struct Response: Decodable {
       let sessionId: String
@@ -81,7 +62,10 @@ public struct WebDriverClient {
       let capabilities: [String: String] = [:]
       let desiredCapabilities: [String: String] = [:]
     }
-    let httpClient: WebDriverHTTPClient = Curl.findExecutable() ?? httpClient
+    let httpClient = httpClient ??
+      CurlWebDriverHTTPClient.find() ??
+      URLSessionWebDriverHTTPClient(session: .shared)
+    
     var request = URLRequest(url: endpoint.appendingPathComponent("session"))
     request.httpMethod = "POST"
     request.httpBody = body.data(using: .utf8)
@@ -125,71 +109,5 @@ public struct WebDriverClient {
     var request = URLRequest(url: URL(string: makeSessionURL())!)
     request.httpMethod = "DELETE"
     _ = try await client.data(for: request)
-  }
-}
-
-
-private protocol WebDriverHTTPClient {
-  func data(for request: URLRequest) async throws -> Data
-}
-
-extension URLSession: WebDriverHTTPClient {
-  func data(for request: URLRequest) async throws -> Data {
-    let (data, httpResponse) = try await self.data(for: request)
-    guard let httpResponse = httpResponse as? HTTPURLResponse, 200..<300 ~= httpResponse.statusCode
-    else {
-      throw WebDriverError.httpError(
-        "\(request.httpMethod ?? "GET") \(request.url.debugDescription) failed"
-      )
-    }
-    return data
-  }
-}
-
-private struct Curl: WebDriverHTTPClient {
-  let cliPath: URL
-
-  static func findExecutable() -> Curl? {
-    guard let path = ProcessInfo.processInfo.environment["PATH"] else { return nil }
-    #if os(Windows)
-    let pathSeparator: Character = ";"
-    #else
-    let pathSeparator: Character = ":"
-    #endif
-    for pathEntry in path.split(separator: pathSeparator) {
-      let candidate = URL(fileURLWithPath: String(pathEntry)).appendingPathComponent("curl")
-      if FileManager.default.fileExists(atPath: candidate.path) {
-        return Curl(cliPath: candidate)
-      }
-    }
-    return nil
-  }
-
-  func data(for request: URLRequest) async throws -> Data {
-    guard let url = request.url?.absoluteString else {
-      preconditionFailure()
-    }
-    let process = Process()
-    process.executableURL = cliPath
-    process.arguments = [
-      url, "-X", request.httpMethod ?? "GET", "--silent", "--fail-with-body", "--data-binary", "@-"
-    ]
-    let stdout = Pipe()
-    let stdin = Pipe()
-    process.standardOutput = stdout
-    process.standardInput = stdin
-    if let httpBody = request.httpBody {
-      try stdin.fileHandleForWriting.write(contentsOf: httpBody)
-    }
-    try stdin.fileHandleForWriting.close()
-    try process.run()
-    process.waitUntilExit()
-    let responseBody = try stdout.fileHandleForReading.readToEnd()
-    guard process.terminationStatus == 0 else {
-      throw WebDriverError.httpError(
-        responseBody.flatMap { String(data: $0, encoding: .utf8) } ?? ""
-      )
-    }
-    return responseBody ?? Data()
   }
 }
